+++ assets/javascripts/kitten/components/videos/embed-player/test.js
@@ -1,17 +1,30 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { EmbedPlayer } from '../../components/media/embed-player'
+import { EmbedPlayer } from './index'
 
 describe('<EmbedPlayer />', () => {
   let component
 
-  describe('by default', () => {
+  describe('with default props', () => {
     beforeEach(() => {
-      component = renderer.create(<EmbedPlayer />).toJSON()
+      component = renderer
+        .create(
+          <EmbedPlayer
+            previewProps={{
+              thumbnail: {
+                src: 'https://placekitten.com/620/376',
+                alt: 'An image',
+              },
+            }}
+            playButtonLabel="Play"
+            ratio={50}
+          />,
+        )
+        .toJSON()
+    })
 
-      it('matches with snapshot', () => {
-        expect(component).toMatchSnapshot()
-      })
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
     })
   })
 
@@ -22,8 +35,8 @@ describe('<EmbedPlayer />', () => {
           <EmbedPlayer
             previewProps={{
               thumbnail: {
-                src: 'https://placehold.it/620x376/caf4fe/caf4fe',
-                alt: 'Une image',
+                src: 'https://placekitten.com/620/376',
+                alt: 'An image',
               },
             }}
             playButtonLabel="Play"