+++ CHANGELOG.md
@@ -12,7 +12,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `TitleWithStroke` component.
 - Feature: Add `titleModifierStyles` common style for titles.
 - Feature: Add `strokeModifierStyles` common style for horizontal strokes.
+- Fix: Modernize `Video` component.
+- Fix: Use `styled-components` instead of `Radium` for `EmbedPlayer` component.
 - Doc: Add a doc for style `utilities`.
+- Feature: Add pagination position `bottom-left` and `bottom-right` on Carousel.
+- Feature: Add pagination squares via `showPageSquares` prop on Carousel.
+- Feature: Add Loop feature and `loop` prop on Carousel.
 
 ## [2.64.0] - 2020-04-07